Split pattern parser into individual modules.
